Renamed AliasesService file
