+++ lib/features/aliases/presentation/controller/aliases_notifier.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:anonaddy/features/aliases/data/alias_service.dart';
+import 'package:anonaddy/features/aliases/data/aliases_service.dart';
 import 'package:anonaddy/features/aliases/domain/alias.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
