+++ lib/features/aliases/presentation/controller/default_recipient/default_recipient_notifier.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:anonaddy/common/utilities.dart';
-import 'package:anonaddy/features/aliases/data/alias_service.dart';
+import 'package:anonaddy/features/aliases/data/aliases_service.dart';
 import 'package:anonaddy/features/aliases/presentation/controller/alias_screen_notifier.dart';
 import 'package:anonaddy/features/aliases/presentation/controller/default_recipient/default_recipient_state.dart';
 import 'package:anonaddy/features/recipients/domain/recipient.dart';
