+++ lib/features/create_alias/presentation/controller/create_alias_notifier.dart
@@ -5,7 +5,7 @@ import 'package:anonaddy/common/constants/toast_message.dart';
 import 'package:anonaddy/common/utilities.dart';
 import 'package:anonaddy/features/account/data/account_service.dart';
 import 'package:anonaddy/features/account/domain/account.dart';
-import 'package:anonaddy/features/aliases/data/alias_service.dart';
+import 'package:anonaddy/features/aliases/data/aliases_service.dart';
 import 'package:anonaddy/features/aliases/presentation/controller/aliases_notifier.dart';
 import 'package:anonaddy/features/create_alias/presentation/controller/create_alias_state.dart';
 import 'package:anonaddy/features/domain_options/data/domain_options_service.dart';
